Adding int casting to sizeof()

+++ cpp/src/arrow/csv/lexing_internal.h
@@ -136,8 +136,8 @@ class SSE42Filter {
     // Look up every byte in `w` in the SIMD filter. Use the explicit-length
     // comparison since `w` may contain an embedded NUL byte, which the
     // implicit-length _mm_cmpistrc would otherwise treat as a terminator.
-    return _mm_cmpestrc(_mm_set1_epi64x(w), sizeof(WordType), filter_,
-                        sizeof(BulkFilterType),
+    return _mm_cmpestrc(_mm_set1_epi64x(w), static_cast<int>(sizeof(WordType)), filter_,
+                        static_cast<int>(sizeof(BulkFilterType)),
                         _SIDD_UBYTE_OPS | _SIDD_CMP_EQUAL_ANY);
   }
